feat(ui): Moved several files to be under (tabs), to display the navbar across home-cabinet-profile
- Updated (tabs)/_layout.tsx
- Updated screens/_layout.tsx
- Removed placeholder BottomBar in (tabs)/home
- Properly navigates between main tabs
- Bug: issue with selecting profile showing public view rather than private view (path issue)
