+++ frontend/app/(tabs)/home.tsx
@@ -1,7 +1,6 @@
 import { View, Text, StyleSheet } from "react-native";
 import { router } from "expo-router";
 import FormButton from "@/components/ui/FormButton";
-import BottomNav from "@/components/ui/BottomNav";
 
 export default function HomeScreen() {
   return (
@@ -10,45 +9,12 @@ export default function HomeScreen() {
         <Text style={styles.title}>Cocktail Mixology</Text>
         <Text style={styles.subtitle}>Welcome! Choose a section:</Text>
 
-      <FormButton
-        title="Scan Ingredients"
-        onPress={() => router.push("/screens/ingredient-scanner")}
-      />
+        <FormButton title="Scan Ingredients" onPress={() => router.push("/screens/ingredient-scanner")} />
+        <FormButton title="My Cabinet" onPress={() => router.push("/my-ingredients")} />
+        <FormButton title="Recommendations" onPress={() => router.push("/screens/recommendations")} />
+        <FormButton title="Profile" onPress={() => router.push("/user-profile")} />
+        <FormButton title="Settings" onPress={() => router.push("/settings")} />
 
-      <FormButton
-        title="My Cabinet"
-        onPress={() => router.push("/screens/my-ingredients")}
-      />
-
-      <FormButton
-        title="Recommendations"
-        onPress={() => router.push("/screens/recommendations")}
-      />
-
-      <FormButton
-        title="Profile"
-        onPress={() => router.push("/screens/user-profile")}
-      />
-
-      <FormButton
-        title="Settings"
-        onPress={() => router.push("/(settings)/settings")}
-      />
-      </View>
-
-      {/* floating dock */}
-      <View style={styles.dock}>
-        <BottomNav
-          activeIndex={0}
-          safeArea={false}
-          items={[
-            { icon: "home-outline" }, // home
-            { icon: "heart-outline" }, //   favorites
-            { icon: "search-outline" }, //  search/browse
-            { icon: "cube-outline", route: "/screens/my-ingredients" }, // Cabinet
-            { icon: "person-outline", route: "/screens/user-profile" },  // Profile
-          ]}
-        />
       </View>
     </View>
   );
